优化输出，增加对输入命令的判断,优化Controller中一些方法的执行,重新建立数据库数据进行测试

+++ src/main/java/ss/controller/ProductController.java
@@ -1,11 +1,14 @@
 package ss.controller;
 
+import ss.bo.QueryStoreBo;
 import ss.constant.Constant;
 import ss.po.Product;
 import ss.po.View;
 import ss.service.ProductService;
+import ss.service.StoreService;
 import ss.service.ViewService;
 import ss.service.impl.ProductServiceImpl;
+import ss.service.impl.StoreServiceImpl;
 import ss.service.impl.ViewServiceImpl;
 import ss.utils.ListUtils;
 import ss.utils.RandomUtils;
@@ -19,6 +22,8 @@ import java.util.List;
 public class ProductController {
 
     static ProductService productService = new ProductServiceImpl();
+
+    static StoreService storeService = new StoreServiceImpl();
 
     static ViewService viewService = new ViewServiceImpl();
 
@@ -34,7 +39,7 @@ public class ProductController {
         if(!viewService.checkUser(view)){
             return "权限不足，请联系root用户进行操作";
         }
-        if(inputList.size()>1 && "ins".equals(inputList.get(1))){
+        if(inputList.size()>1 && Constant.INSERT_ARG.equals(inputList.get(1))){
             String productName = null, description = null;
             for(int i = 2;i<inputList.size();i += 2){
                 if("-n".equals(inputList.get(i))){
@@ -51,7 +56,7 @@ public class ProductController {
                 return "新商品插入成功";
             }
         }
-        if(inputList.size()>1 && "alter".equals(inputList.get(1))){
+        if(inputList.size()>1 && Constant.ALTER_ARG.equals(inputList.get(1))){
             Product product = new Product();
             for(int i = 2;i<inputList.size();i += 2){
                 if("-i".equals(inputList.get(i))){
@@ -70,6 +75,9 @@ public class ProductController {
             return "商品修改失败";
         }
         if(inputList.size()>1 && "-d".equals(inputList.get(1))){
+            if(!storeService.queryStoreByQueryBo(new QueryStoreBo(null, inputList.get(2))).isEmpty()){
+                return "仍有超市存在该商品库存,不可删除该商品";
+            }
             if(productService.deleteProduct(inputList.get(2))){
                 return "商品删除成功";
             }
